pyproject.toml
[project]
name = "the-scrapper"
version = "0.1.0"
description = "Flow-centric, plugin-based scraping framework (async-native) with schema validation and SQLite sink."
requires-python = ">=3.11"
readme = "README.md"
license = { text = "MIT" }
authors = [{ name = "The Scrapper Contributors" }]
dependencies = [
  "httpx>=0.27.0",
  "requests>=2.31.0",
  "beautifulsoup4>=4.12.3",
  "lxml>=5.2.2",
  "pydantic>=2.7.0",
  "PyYAML>=6.0.1",
  "prometheus-client>=0.20.0",
]

[project.optional-dependencies]
browser = [
  "playwright>=1.45.0",
]

[project.scripts]
scrape = "scrapper_cli.main:main"

[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["scrapper", "scrapper_cli", "plugins"]
